requirements.txt
# Core AI Framework
langchain==1.2.10
langchain-groq==1.1.2
langchain-pinecone==0.2.13
langchain-community==0.4.1
langchain_huggingface
streamlit-lottie

# Document Processing
pypdf==6.7.0

# Infrastructure & Utilities
python-dotenv==1.2.1
pinecone-client==6.0.0
groq==0.37.1

# Web Server (For Step 2 of Day 2)
fastapi==0.115.0
uvicorn==0.30.0